utils/pdb_utils.py
# utils/pdb_utils.py

"""
Utilities for PDB handling: structure prediction, docking file parsing, etc.
Used by the docking_module.
"""

import os
import logging
from typing import Tuple, List, Optional

logger = logging.getLogger("G-Synth:PDBUTILS")

# Attempt imports for structure prediction & docking
try:
    from transformers import AutoModelForCausalLM, AutoTokenizer
    import torch
    USING_ESMFOLD = True
except ImportError:
    USING_ESMFOLD = False
    logger.warning("ESMFold (Transformers) not available. Structure prediction disabled.")

try:
    import Bio.PDB as PDB
    USING_BIOPYTHON_PDB = True
except ImportError:
    USING_BIOPYTHON_PDB = False
    logger.warning("Biopython PDB not available. Some PDB parsing features disabled.")

def predict_protein_structure(sequence: str, out_pdb_path: str) -> Optional[str]:
    """
    Predict protein structure from sequence using ESMFold (if available).
    Saves PDB to out_pdb_path. Returns path if successful, else None.
    """
    if not USING_ESMFOLD:
        logger.error("ESMFold not available.")
        return None
    try:
        # NOTE: This is a placeholder. Actual ESMFold from Hugging Face requires specialized pipelines.
        # Here we assume existence of an imaginary function `esmfold_predict`.
        from transformers import HfModelHub  # hypothetical
        # ... pseudo-code for ESMFold inference ...
        # model = EsmFoldModel.from_pretrained("facebook/esmfold_v1")
        # pdb_str = model.predict_pdb(sequence)
        # with open(out_pdb_path, "w") as f:
        #     f.write(pdb_str)
        # For demonstration, we simply copy a placeholder or fail.
        logger.info(f"ESMFold prediction requested for sequence of length {len(sequence)}.")
        return None
    except Exception as e:
        logger.error(f"Error in ESMFold prediction: {e}")
        return None

def build_bform_dna_model(sequence: str, out_pdb_path: str) -> Optional[str]:
    """
    Build simplistic B-form DNA PDB from sequence. Requires Biopython.
    Returns PDB path or None.
    """
    if not USING_BIOPYTHON_PDB:
        logger.error("Biopython PDB not available.")
        return None
    try:
        # Placeholder: full implementation is lengthy; normally would use Bio.PDB to build base pairs.
        logger.info(f"Building B-form DNA model for length {len(sequence)}.")
        return None
    except Exception as e:
        logger.error(f"Error building B-form DNA: {e}")
        return None

def parse_interface_residues(pdb1: str, pdb2: str, complex_pdb: str, cutoff: float = 5.0) -> List[Tuple[str, str, int]]:
    """
    Identify residues within cutoff (Å) between two chains in a docked complex.
    Returns list of tuples: (chain1_resname, chain2_resname, residue_number).
    """
    if not USING_BIOPYTHON_PDB:
        return []
    try:
        parser = PDB.PDBParser(QUIET=True)
        structure = parser.get_structure("complex", complex_pdb)
        # Simplified: assume two chains A and B
        model = structure[0]
        chainA = model[list(model.child_dict.keys())[0]]
        chainB = model[list(model.child_dict.keys())[1]]  # second chain
        ns = PDB.NeighborSearch([atom for atom in structure.get_atoms()])
        interface = []
        for resA in chainA:
            for atomA in resA:
                close = ns.search(atomA.coord, cutoff, level="R")  # search residues within cutoff
                for resB in close:
                    if resB.get_parent().id == chainB.id:
                        interface.append((resA.get_resname(), resB.get_resname(), resA.id[1]))
        return list(set(interface))
    except Exception as e:
        logger.error(f"Error parsing interface: {e}")
        return []

####################################################################
# End of pdb_utils.py
####################################################################
